Rename Navigation folder to FindUsages
And update namespaces
