+++ unity/EditorPlugin/SinceUnity.5.6/FindUsages/Window/AnimatorElement.cs
@@ -2,21 +2,21 @@ using System;
 using JetBrains.Annotations;
 using JetBrains.Rider.Model.Unity.BackendUnity;
 
-namespace JetBrains.Rider.Unity.Editor.Navigation.Window
+namespace JetBrains.Rider.Unity.Editor.FindUsages.Window
 {
     [Serializable]
     internal class AnimatorElement : AbstractUsageElement
     {
         [NotNull]
         public const string AnimatorStateMachineIcon = "AnimatorStateMachine Icon";
-        
+
         [NotNull]
         public const string AnimatorStateIcon = "AnimatorState Icon";
-        
+
         [NotNull] public string[] PathElements { get; }
         public AnimatorUsageType Type { get; }
-        
-        
+
+
         public AnimatorElement(AnimatorUsageType type,
                                string filePath,
                                string fileName,
