+++ unity/EditorPlugin/SinceUnity.5.6/FindUsages/Window/PrefabElement.cs
@@ -1,11 +1,12 @@
 using System;
 
-namespace JetBrains.Rider.Unity.Editor.Navigation.Window
+namespace JetBrains.Rider.Unity.Editor.FindUsages.Window
 {
   [Serializable]
   internal class PrefabElement : AbstractUsageElement
   {
-    public PrefabElement(string filePath, string fileName, string[] path, int[] rootIndices) : base(filePath, fileName, path, rootIndices)
+    public PrefabElement(string filePath, string fileName, string[] path, int[] rootIndices)
+      : base(filePath, fileName, path, rootIndices)
     {
     }
